WikiBuddy2.0/Wikibuddy/requirements.txt
# requirements.txt

# Async HTTP requests
aiohttp==3.8.4

# HTML parsing
beautifulsoup4==4.11.2

# Vector database
faiss-cpu==1.7.3

# Numerical computations
numpy==1.23.5

# Deep learning and NLP
torch==1.13.1
transformers==4.26.1

# Async programming
asyncio==3.4.3

# Data classes
dataclasses==0.6

# URL handling
requests==2.28.2
